Bug fixes
Workspace variables weren't being saved properly because the saveResults function doesn't have access to the global workspace. For now, the code from saveResults.m is in the SettlingTubeInterface.m file. For a cleaner interface script, the data and figures should be packaged as structs and passed to the saveResults function so that it can explicitly save the workspace variables of interest.