Cleaned up functions
Removed unnecessary functions. The functions used to find the mode of the data using complicated loops were relict artifacts of QuickBasic. The mode is now found using MATLAB functions. The reporting of mode is also greatly improved because only one value is reported, unlike the old code structure, which could report many values and required data review to determine which was the true mode.